disable successRedirect and session
* There will be no session cookie generated by passportjs returned to the
client
* Redirection to home page is not needed and will continue excution in the
next middleware of the same route /dmm-account
* POST /login renamed to /dmm-account
