+++ src/views/index.hbs
@@ -1,6 +1,6 @@
 <h1>Modcolle - Kancolle Proxy Server<h1>
 
-<form action="/login" method="POST">
+<form action="/dmm-account" method="POST">
   <fieldset>
     <legend>Login by DMM Account</legend>
     <label>Email:</br>
